Add currency to Drawer Nav and handle currency page
