+++ ECommerceApp/App.js
@@ -9,7 +9,7 @@ import WishList from './screens/WishList';
 import Cart from './screens/Cart';
 import SettingProfile from './screens/SettingProfile';
 import Payment from './screens/Payment';
-import Page1 from './screens/Page1';
+
 import Address from './screens/Address';
 import NewAddress from './screens/NewAddress';
 import Drawerheader from './components/Drawerheader';
